Move output formats to separate file than models
