+++ cdse_covid/semantic_extraction/models/amr.py
@@ -1,23 +1,14 @@
 """Classes related to AMR Model."""
-from dataclasses import dataclass
 import logging
 from os import chdir, getcwd
 from pathlib import Path
-from typing import Any, List
+from typing import List
 import uuid
 
 from amr_utils.amr_readers import AMR_Reader, Metadata_Parser  # pylint: disable=import-error
 from transition_amr_parser.parse import AMRParser  # pylint: disable=import-error
 
-
-@dataclass
-class AMROutput:
-    """Class to hold AMR Output."""
-
-    label_id: int
-    graph: Any
-    alignments: Any
-    annotations: Any
+from cdse_covid.semantic_extraction.models.output_formats import AMROutput
 
 
 class AMRModel(object):
